Refactor ZBQueue for Localization support
ZBQueue was using the strings "Unresolved Dependencies" and
"Conflictions" to capture state, which got mixed in with string
representations of ZBQueueType. These strings were also used in the
UI, which made Localization impossible.

I added ZBQueueTypeUnresolvedDependencies and ZBQueueTypeConflictions
as new ZBQueueTypes, and which got rid of the need to use strings to
capture state. Now these support Localization, plus it's a bit cleaner
and faster since it uses ZBQueueType now instead of comparing strings.

I also went ahead and added ZBQueueTypeUnknown and replaced that
wherever "0" was used, so it's a bit better documented. Plus some
macros for looping ZBQueueTypes. And removed -[ZBQueue keyToQueue].