Fixed some of the localizations strings.
The first translation has had some inconsistencies that didn't match with the official Apple translations for 'restoring' etc. Moreover, the line 276 didn't have any sense, because of the wrong translation of the word 'depends' into Polish. Finally, the change in line 217 was made, because of the UX, as every search engine uses the new phrase rather than the earlier one.