Fix queue navigation bar color bug
Queue navigation bar would stay dark after installing packages because it was set to dark mode and never set back. The simplest solution is to create a new navigation controller each time and release it as needed, instead of trying to manage a back-and-forth state of the same navigation controller.